Move all components logic to custom hooks
