+++ src/app/src/components/Alerts.jsx
@@ -1,51 +1,16 @@
-import { ALERT_TYPES, SUCCESS_ALERT_HIDE_MS, ERROR_ALERT_HIDE_MS } from "../utils/constants";
-import { selectAlerts, clearAlert } from "../store/slices/alertsSlice";
-
-import { Snackbar, Alert} from "@mui/material";
-import { useSelector, useDispatch } from "react-redux";
-import { useState } from "react";
-import { useEffect } from "react";
-
+import { Snackbar} from "@mui/material";
+import useAlerts from "../utils/hooks/useAlerts";
 
 const Alerts = () => {
-    const [show, setShow] = useState(false);
-    const alert = useSelector(selectAlerts);
-    const dispatch = useDispatch();
+    const {
+      show,
+      autoHideDuration,
+      handleAlertClose,
+      alertComponent
+    } = useAlerts();
 
-    useEffect(() => {
-      if (!alert || alert.length === 0) {
-        setShow(false);
-        return;  
-      }
-
-      setShow(true);
-    }, [show, alert]);
-
-    const handleAlertClose = () => {
-      dispatch(clearAlert());
-    }
-
-    if (!alert) {
-        return null;
-    }
-
-    const {type, message} = alert;
-    let alertComponent = null;
-    let autoHideDuration = SUCCESS_ALERT_HIDE_MS;
-    switch(type) {
-        case ALERT_TYPES.info:
-            alertComponent = <Alert severity="info">{message}</Alert>;
-            break;
-        case ALERT_TYPES.warn:
-            alertComponent = <Alert severity="warning">{message}</Alert>;
-            break;
-        case ALERT_TYPES.success:
-            alertComponent = <Alert severity="success">{message}</Alert>;  
-            break;
-        case ALERT_TYPES.error:
-            alertComponent = <Alert severity="error">{message}</Alert>;
-            autoHideDuration = ERROR_ALERT_HIDE_MS;
-            break;
+    if (!show) {
+      return null;
     }
 
     return (
