+++ src/app/src/pages/ProtectedPage.jsx
@@ -1,21 +1,7 @@
-import navigation from '../utils/navigation';
-import { selectIsUserAuth } from '../store/slices/authSlice';
-
-import { useNavigate } from "react-router-dom"
-import { useEffect } from 'react';
-import {useSelector} from "react-redux";
+import useProtectedPage from "../utils/hooks/useProtectedPage";
 
 const ProtectedPage = ({children}) => {
-	const isLoggedIn = useSelector(selectIsUserAuth);
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (!isLoggedIn) {
-      // TODO: Dispatch an alert about log in
-
-      navigate(navigation.login, {replace: true});
-    }
-  }, [isLoggedIn, navigate]);
+  useProtectedPage();
 
   return (
     children
